WIP refactoring envs
