+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: eduribei <eduribei@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 20:52:30 by eduribei          #+#    #+#             */
-/*   Updated: 2025/03/21 22:21:24 by eduribei         ###   ########.fr       */
+/*   Updated: 2025/03/22 11:37:57 by eduribei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 #include <stdlib.h>						// for malloc etc
 #include <readline/readline.h>			// for readline
 #include <readline/history.h>			// for history
+
 
 typedef enum e_tok_exit
 {
